pyproject.toml
[project]
name = "vinted-notifier"
version = "0.1.0"
description = "Receive Apprise notifications when new Vinted items matching your criteria are added."
readme = "README.md"
authors = [
  { name = "cchrkk" }
]
requires-python = ">=3.10"
license = { text = "MIT" }
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]
dependencies = [
  "requests>=2.30.0",
  "PyYAML>=6.0",
  "apprise>=1.8.0"
]

[project.scripts]
vinted-notifier = "vinted_notifier.cli:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
